fix: log memory usage not working
